[refactor] Home Entity 리팩토링
